+++ src/App.js
@@ -1,12 +1,12 @@
 import './App.css';
 import axios from 'axios';
 import { useState } from 'react';
-import DisplayTrivia from './DisplayTrivia.js';
-import SaveButton from './SaveButton.js';
+import {BrowserRouter as Router, Route} from 'react-router-dom';
+import SavedGames from './SavedGames.js';
+import SelectTrivia from './SelectTrivia';
 
 
 function App() {
-  const [answerCheck, setAnswerCheck] = useState();
   const [userInput, setUserInput] = useState({
     categories: "placeholder",
     questionNum: "placeholder"
@@ -25,73 +25,26 @@ function App() {
     }).then( (response) => {
       setUserQuestion(response.data.results)
     })
-  }
-
-  const handleSubmit = (e) => {
-    e.preventDefault(e);
-    getTrivia();
   };
 
-  const handleChange = (e) => {
-    const value = e.target.value;
-    setUserInput({
-      ...userInput,
-      [e.target.name]: value
-    })
-  }
-
   return (
-    <div className="App">
-      <form action="submit" onSubmit={handleSubmit}>
-        <select name="categories" id="categories" value={userInput.categories} onChange={handleChange}>
-          <option value="placeholder" disabled>Please select your categories</option>
-          <option name="General Knowledge" value="9">General Knowledge</option>
-          <option name="Sports" value="21">Sports</option>
-          <option name="Geography" value="22">Geography</option>
-          <option name="Celebrities" value="26">Celebrities</option>
-          <option name="Vehicles" value="28">Vehicles</option>
-          <option name="Entertainment: Film" value="11">Film</option>
-          <option name="Entertainment: Japanese Anime & Manga" value="31">Anime and Manga</option>
-        </select>
+    <Router>
+      <div className="App">
 
-        <select name="questionNum" id="questionNum" value={userInput.questionNum} onChange={handleChange}>
-          <option value="placeholder" disabled>Select the amount of questions</option>
-          <option name="5" value="5">5</option>
-          <option name="6" value="6">6</option>
-          <option name="7"value="7">7</option>
-          <option name="8"value="8">8</option>
-          <option name="9"value="9">9</option>
-          <option name="10"value="10">10</option>
-        </select>
+        <Route exact path="/" render={() =>
+          <SelectTrivia
+            getTrivia={getTrivia}
+            userQuestion={userQuestion}
+            setUserQuestion={setUserQuestion}
+            userInput={userInput}
+            setUserInput={setUserInput}
+          />
+        }/>
 
-        <button type="submit">Start Trivia</button>
-      </form>
-      {
-        userQuestion.map((key, i) => {
-          return(
-            <DisplayTrivia
-              category={key.category}
-              correctAnswer={key.correct_answer}
-              incorrectAnswer={key.incorrect_answers}
-              question={key.question}
-              key={`key${i}`}
-              answerChecker={setAnswerCheck}
-            />
-          )
-        })
-      }
-      {
-        answerCheck === true ? (
-          <div>
-            <h2>You are correct!</h2>
-          </div>
-        ) : answerCheck === false ? (
-          <h2>You are wrong!</h2>
-        ) : null
-      }
-
-      <SaveButton currentTrivia={userQuestion}/>
-    </div>
+        <Route exact path="/savedGames" component={SavedGames}/>
+        
+      </div>
+    </Router>
   );
 }
