+++ src/SavedGames.js
@@ -1,0 +1,15 @@
+import {Link} from 'react-router-dom';
+import SelecTrivia from './SelectTrivia.js'
+
+const SavedGames = () => {
+    return(
+        <div>
+            <button>
+              <Link to="/">Current Game</Link>
+            </button>
+            <h2>hello</h2>
+        </div>
+    )
+}
+
+export default SavedGames;
